Some minor improvements all over the place.
Fix issue that caused the style intead of the size property to be used for font scaling in the MapObjectInspector.
