+++ utiliti/src/de/gurkenlabs/utiliti/swing/panels/MapObjectInspector.java
@@ -130,11 +130,11 @@ public class MapObjectInspector extends PropertyPanel implements PropertyInspect
 
     this.lblRendering = new JLabel("");
     this.lblRendering.setForeground(Color.GRAY);
-    this.lblRendering.setFont(lblRendering.getFont().deriveFont(lblRendering.getFont().getStyle() * 0.75f));
+    this.lblRendering.setFont(lblRendering.getFont().deriveFont(lblRendering.getFont().getSize() * 0.75f));
 
     this.lblLayer.setHorizontalAlignment(SwingConstants.TRAILING);
     this.lblLayer.setForeground(Color.GRAY);
-    this.lblLayer.setFont(this.lblLayer.getFont().deriveFont(lblLayer.getFont().getStyle() * 0.75f));
+    this.lblLayer.setFont(this.lblLayer.getFont().deriveFont(lblLayer.getFont().getSize() * 0.75f));
 
     panel_1.add(lblEntityId);
     panel_1.add(Box.createHorizontalStrut(47));
